Bugfixes: get the app to compile
I couldn't run the migrations or start the app until I fixed these
things.

* `_MODULE_` to `__MODULE__`
* `con` to `conn`
* `model` to `struct` (the argument to the function is called `struct`,
so `model` is undefined)

Also, note that I removed the second array that you were passing to the
`cast` methods. As we discussed in class a couple times, the book is
outdated. It indicates that you pass an array of required fields, and a
separate array of optional fields. This is no longer true. The new
convention is also to use an array of atoms (`[:name, :username]`)
instead of an array of strings (`~w(name username)`, which is equivalent
to `["name", "username"]`).